AuthenticationAndAuthorization

+++ Restaurant.Application/Dishes/Commands/DeleteAllByRestaurantId/DeleteAllByRestaurantIdCommandHandler.cs
@@ -3,13 +3,14 @@ using MediatR;
 using Microsoft.Extensions.Logging;
 using Restaurants.Domain.Entities;
 using Restaurants.Domain.Exceptions;
+using Restaurants.Domain.Interfaces;
 using Restaurants.Domain.Repositories;
 
 namespace Restaurants.Application.Dishes.Commands.DeleteAllByRestaurantId;
 
 public class DeleteAllByRestaurantIdCommandHandler(IDishRepository dishRepository,
     IRestaurantRepository restaurantRepository,
-    IMapper mapper,
+    IRestaurantAuthorizationService restaurantAuthorizationService,
     ILogger<DeleteAllByRestaurantIdCommandHandler> logger) : IRequestHandler<DeleteAllByRestaurantIdCommand>
 {
     public async Task Handle(DeleteAllByRestaurantIdCommand request, CancellationToken cancellationToken)
@@ -18,6 +19,9 @@ public class DeleteAllByRestaurantIdCommandHandler(IDishRepository dishRepositor
 
         var restaurant = await restaurantRepository.GetByIdAsync(request.RestaurantId)
             ?? throw new NotFoundException<Restaurant>(request.RestaurantId.ToString());
+
+        if (!restaurantAuthorizationService.Authorize(restaurant, Domain.Constants.ResourceOperation.Update))
+            throw new ForbidException<Restaurant>(restaurant.Id.ToString());
 
         await dishRepository.Delete(restaurant.Dishes);
     }
